third_party/blink/web_tests/external/wpt/css/css-values/exp-log-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'exp(0)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'exp(0)' should be valid in opacity. got disallowed value ""
FAIL 'scale(exp(0))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(exp(0))' should be valid in transform. got disallowed value ""
PASS 'exp(0)' as a computed value should serialize as '1'.
FAIL 'scale(exp(0))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(exp(0))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'log(1)' as a specified value should serialize as 'calc(0)'. assert_not_equals: 'log(1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(log(1))' as a specified value should serialize as 'scale(calc(0))'. assert_not_equals: 'scale(log(1))' should be valid in transform. got disallowed value ""
FAIL 'log(1)' as a computed value should serialize as '0'. assert_equals: 'log(1)' and '0' should serialize the same in computed values. expected "0" but got "1"
FAIL 'scale(log(1))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'. assert_equals: 'scale(log(1))' and 'matrix(0, 0, 0, 0, 0, 0)' should serialize the same in computed values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
FAIL 'calc(exp(0) + log(1) + log(1))' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'calc(exp(0) + log(1) + log(1))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(exp(0) + log(1) + log(1)))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(calc(exp(0) + log(1) + log(1)))' should be valid in transform. got disallowed value ""
PASS 'calc(exp(0) + log(1) + log(1))' as a computed value should serialize as '1'.
FAIL 'scale(calc(exp(0) + log(1) + log(1)))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(calc(exp(0) + log(1) + log(1)))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'calc(log(1) + 0.5)' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(log(1) + 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(log(1) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(log(1) + 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(log(1) + 0.5)' as a computed value should serialize as '0.5'. assert_equals: 'calc(log(1) + 0.5)' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(log(1) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(log(1) + 0.5))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(log(0))' as a specified value should serialize as 'calc(-infinity)'. assert_not_equals: 'calc(log(0))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(log(0)))' as a specified value should serialize as 'scale(calc(-infinity))'. assert_not_equals: 'scale(calc(log(0)))' should be valid in transform. got disallowed value ""
FAIL 'calc(log(0))' as a computed value should serialize as '-infinity'. assert_equals: '-infinity' should round-trip exactly in computed values. expected "-infinity" but got "1"
FAIL 'scale(calc(log(0)))' as a computed value should serialize as 'matrix(-infinity, 0, 0, -infinity, 0, 0)'. assert_equals: 'matrix(-infinity, 0, 0, -infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(-infinity, 0, 0, -infinity, 0, 0)" but got "none"
Harness: the test ran to completion.
